// src/app/modules/warehouse/models/Warehouse.ts
export class Warehouse {
    id?: number;
    name?: any;
    code?: any;
    email?: any;
    contactname?: any;
    addressline1?: any
    addressline2?: any;
    addressline3?: any;
    city?: any;
    state?: any;
    country?: any;
    postcode?: any;
    defaultWarehouse?: any;
    createdDate?: any;
}