// src/app/modules/warehouse/models/WarehouseLocation.ts
export class WarehouseLocation {
    id?: number;
    aisle?: any;
    aisleList?: any;
    bay?: any;
    bayList?: any;
    bin?: any;
    binList?: any;
    defaultLocation?: any;
    kartzhubUserId?: any;
    location?: any;
    modifiedDate?: any;
    quantity?: any;
    shelf?: any;
    shelfList?: any;
    warehouseCode?: any;
    warehouseId?: any; 
    warehouseName?: any;
}